landing page
I need assistance. I left the DARC logo in its original format because it is not showing up on the website. Now that I changed the button it's also not showing. Maybe I did something wrong? I tried looking it up on google, but it showed me the html and body tags